requirements.txt
pyttsx3
SpeechRecognition
pywhatkit
wikipedia
requests
python-decouple
